+++ packages/trackback-agent/CHANGELOG.md
@@ -3,6 +3,206 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.1-alpha.35](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.34...@trackback/agent@0.0.1-alpha.35) (2021-11-21)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.34](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.33...@trackback/agent@0.0.1-alpha.34) (2021-11-18)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.33](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.32...@trackback/agent@0.0.1-alpha.33) (2021-11-18)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.32](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.31...@trackback/agent@0.0.1-alpha.32) (2021-11-18)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.31](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.30...@trackback/agent@0.0.1-alpha.31) (2021-11-18)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.30](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.29...@trackback/agent@0.0.1-alpha.30) (2021-11-18)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.29](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.28...@trackback/agent@0.0.1-alpha.29) (2021-11-17)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.28](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.27...@trackback/agent@0.0.1-alpha.28) (2021-11-17)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.27](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.26...@trackback/agent@0.0.1-alpha.27) (2021-11-17)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.26](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.25...@trackback/agent@0.0.1-alpha.26) (2021-11-16)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.25](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.24...@trackback/agent@0.0.1-alpha.25) (2021-11-15)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.24](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.23...@trackback/agent@0.0.1-alpha.24) (2021-11-11)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.23](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.22...@trackback/agent@0.0.1-alpha.23) (2021-11-09)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.22](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.21...@trackback/agent@0.0.1-alpha.22) (2021-11-08)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.21](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.20...@trackback/agent@0.0.1-alpha.21) (2021-11-08)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.20](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.19...@trackback/agent@0.0.1-alpha.20) (2021-11-01)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.19](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.18...@trackback/agent@0.0.1-alpha.19) (2021-11-01)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.18](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.17...@trackback/agent@0.0.1-alpha.18) (2021-10-29)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.17](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.16...@trackback/agent@0.0.1-alpha.17) (2021-10-28)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.16](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.15...@trackback/agent@0.0.1-alpha.16) (2021-10-28)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.15](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.14...@trackback/agent@0.0.1-alpha.15) (2021-10-28)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.14](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.13...@trackback/agent@0.0.1-alpha.14) (2021-10-28)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.13](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.12...@trackback/agent@0.0.1-alpha.13) (2021-10-28)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.12](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.11...@trackback/agent@0.0.1-alpha.12) (2021-10-28)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
+## [0.0.1-alpha.11](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.10...@trackback/agent@0.0.1-alpha.11) (2021-10-28)
+
+**Note:** Version bump only for package @trackback/agent
+
+
+
+
+
 ## [0.0.1-alpha.10](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/agent@0.0.1-alpha.9...@trackback/agent@0.0.1-alpha.10) (2021-10-28)
 
 **Note:** Version bump only for package @trackback/agent