+++ packages/trackback-key/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.1-alpha.12](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/key@0.0.1-alpha.11...@trackback/key@0.0.1-alpha.12) (2021-11-18)
+
+**Note:** Version bump only for package @trackback/key
+
+
+
+
+
+## [0.0.1-alpha.11](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/key@0.0.1-alpha.10...@trackback/key@0.0.1-alpha.11) (2021-11-18)
+
+**Note:** Version bump only for package @trackback/key
+
+
+
+
+
+## [0.0.1-alpha.10](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/key@0.0.1-alpha.9...@trackback/key@0.0.1-alpha.10) (2021-11-18)
+
+**Note:** Version bump only for package @trackback/key
+
+
+
+
+
+## [0.0.1-alpha.9](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/key@0.0.1-alpha.8...@trackback/key@0.0.1-alpha.9) (2021-11-18)
+
+**Note:** Version bump only for package @trackback/key
+
+
+
+
+
+## [0.0.1-alpha.8](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/key@0.0.1-alpha.7...@trackback/key@0.0.1-alpha.8) (2021-11-17)
+
+**Note:** Version bump only for package @trackback/key
+
+
+
+
+
+## [0.0.1-alpha.7](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/key@0.0.1-alpha.6...@trackback/key@0.0.1-alpha.7) (2021-11-16)
+
+**Note:** Version bump only for package @trackback/key
+
+
+
+
+
+## [0.0.1-alpha.6](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/key@0.0.1-alpha.5...@trackback/key@0.0.1-alpha.6) (2021-11-15)
+
+**Note:** Version bump only for package @trackback/key
+
+
+
+
+
+## [0.0.1-alpha.5](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/key@0.0.1-alpha.4...@trackback/key@0.0.1-alpha.5) (2021-10-28)
+
+**Note:** Version bump only for package @trackback/key
+
+
+
+
+
 ## [0.0.1-alpha.4](https://github.com/trackback-blockchain/trackback-verifiable/compare/@trackback/key@0.0.1-alpha.3...@trackback/key@0.0.1-alpha.4) (2021-10-28)
 
 **Note:** Version bump only for package @trackback/key